backend/requirements.txt
Django
celery
psycopg2
pika
djangorestframework
django-cors-headers
django-celery-beat
django-celery-results
channels